// src/app/interface/frames-interface.ts
export interface Frame{
    id: number;
    frame_detail: string;
    frame_moredetail: string;
    frame_timeAt1_start: string;
    frame_timeAt1_end: string;
    frame_timeAt2_start: string;
    frame_timeAt2_end: string;
    frame_timeAt3_start: string;
    frame_timeAt3_end: string;
    frame_timeAt4_start: string;
    frame_timeAt4_end: string;
    frame_timeAt5_start: string;
    frame_timeAt5_end: string;
    frame_timeStrat_start: string;
    frame_timeStrat_end: string;
    frame_timeOp_start: string;
    frame_timeOp_end: string;
    frame_timedetail: string;
    frame_confidentiality: string;
    frame_responsable: string;
}